my_patronus_env/requirements.txt
openai
PyQt5
pillow